move boneslost calc to activity
